requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
nltk>=3.8
numpy>=1.24.0
scikit-learn>=1.3.0
nomic>=1.0.0
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-ollama>=0.1.0
langchain-text-splitters>=0.0.1
langchain-experimental>=0.0.50
langchain-core>=0.1.0
langgraph>=0.2.0
langsmith>=0.1.0
httpx>=0.27.0
mcp>=1.0.0
pyyaml>=6.0
tiktoken>=0.5.0
requests>=2.31.0
guardrails-ai>=0.4.0
pydeid
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
boto3>=1.34.0
pytest-asyncio>=0.23.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
sendgrid>=6.11.0
python-multipart>=0.0.6
email-validator>=2.0.0
psycopg2-binary>=2.9.0
